Add `DomainEventDispatcher` to contain logic for dispatching events
This interface's method will dispatch all events on a set of
entities (via Mediatr's `Publish` method) and remove them afterward.

It is expected that consumers of this library will wire this up via
their dependency injection framework, and call it from an
appropriate place, such as when entities are persisted via
Entity Framework Core's `SaveChangesAsync` method.