Provide our own IRepository with projection features for EF Core
This duplicates the repository methods in the sample IRepository
provided by Ardalis.Specification, but adds new ones that integrate
with AutoMapper to provide projection support.  This allow
developers to separate the concepts of query logic and projection
for more flexibility when querying data.

* FirstOrDefaultProjectedAsync
* SingleOrDefaultProjectedAsync
* ListProjectedAsync

Each of these methods will return a result mapped from the entity
contained within the repository being used.  It requires that an
AutoMapper mapping be configured from the entity class to the
destination class (DTO).

A suite of integration tests has been added for testing the
new methods.  This requires localdb or a containerized instance of
SQL Server.  See the README for instructions.

+semver: breaking