Prevent compiler warning about BaseEntity.Id being null
This will initialize the Id property to the default value for its
type (0 for int, etc.) upon construction.  When persisting via
EF Core this will get set to the database-generated value.  Or, if
you are using a type like Guid for the Id, then set it via the
derived entity's constructor.