Rename to Centeva.DomainModeling
The term "Shared Kernel" is misapplied here.  It is meant to denote
a subset of the domain model of an application or system that is
shared between multiple bounded contexts.  This package is designed
to provide basic building blocks for creating a rich domain model
within your application.

Hence, the rename to Centeva.DomainModeling.*.  It is understood
that some developers are not familiar with the use of "Domain" here,
but it's the most appropriate term I could find so far.

"DataAccess" was also proposed, but this has a subtly different
meaning.  The tools provided by this library are not solely for
providing data access tools, but also to aid in the development of
a domain model using Domain Driven Design terms and patterns.