Refactor domain event to reduce method accessibility
I replaced the IDomainWithEvents interface with an abstract base
class.  This enables us to tighten up method accessibility for
creation and removal of domain events.  I eliminated the
RemoveDomainEvent method, and encourage other ways to deal with
this such as using domain services to orchestrate sequential
operations with a single event dispatched at the end.

This retains the ability to mix different entity Id datatypes in a
single application while being able to gather up domain events for
dispatch as part of the EF Core persistence process.

+semver:breaking