Remove hard dependency on MediatR
This splits the MediatR-backed implementation of domain events into
a separate package, and introduces a parallel implementation that
uses the very similar Mediator library which uses source generation
to avoid runtime reflection.  (And it's open source and free to
use.)

This is a breaking change, in that consuming application will need
to choose a domain events implementation, import the desired
package, and update `using` statements accordingly.  If migrating
to the Mediator implementation, note that event handlers have a
slightly different API (they return ValueTask instead of Task.)

+semver:breaking